+++ src/lesson7/BreadthFirstPath.java
@@ -2,18 +2,11 @@ package lesson7;
 
 import java.util.LinkedList;
 
-public class BreadthFirstPath {
-    private boolean[] marked;
-    private int[] edgeTo;
-    private int source;
-
+public class BreadthFirstPath extends ProtoPaths {
 
     public BreadthFirstPath(Graph g, int source) {
-        this.source = source;
-        marked = new boolean[g.getVertexCount()];
-        edgeTo = new int[g.getVertexCount()];
-
-        bfs(g, source);
+       super(g, source);
+       bfs(g, source);
     }
 
     private void bfs(Graph g, int source) {
@@ -31,23 +24,5 @@ public class BreadthFirstPath {
                 }
             }
         }
-    }
-
-
-    public boolean hasPathTo(int v) {
-        return marked[v];
-    }
-
-    public LinkedList<Integer> pathTo(int v) {
-        if (!hasPathTo(v)) {
-            return null;
-        }
-        LinkedList<Integer> stack = new LinkedList<>();
-        int vertex = v;
-        while (vertex != source) {
-            stack.push(vertex);
-            vertex = edgeTo[vertex];
-        }
-        return stack;
     }
 }
