+++ src/lesson7/DepthFirstPaths.java
@@ -2,15 +2,10 @@ package lesson7;
 
 import java.util.LinkedList;
 
-public class DepthFirstPaths {
-    private boolean[] marked;
-    private int[] edgeTo;
-    private int source;
+public class DepthFirstPaths extends ProtoPaths {
 
     public DepthFirstPaths(Graph g, int source) {
-        this.source = source;
-        marked = new boolean[g.getVertexCount()];
-        edgeTo = new int[g.getVertexCount()];
+        super(g, source);
 
         dfs(g, source);
     }
@@ -24,22 +19,4 @@ public class DepthFirstPaths {
             }
         }
     }
-
-    public boolean hasPathTo(int v) {
-        return marked[v];
-    }
-
-    public LinkedList<Integer> pathTo(int v) {
-        if (!hasPathTo(v)) {
-            return null;
-        }
-        LinkedList<Integer> stack = new LinkedList<>();
-        int vertex = v;
-        while(vertex != source){
-            stack.push(vertex);
-            vertex = edgeTo[vertex];
-        }
-        return stack;
-    }
-
 }
