+++ src/lesson7/ProtoPaths.java
@@ -1,0 +1,33 @@
+package lesson7;
+
+import java.util.LinkedList;
+
+public class ProtoPaths {
+
+    protected boolean[] marked;
+    protected int[] edgeTo;
+    private int source;
+
+    public ProtoPaths(Graph g, int source) {
+        this.source = source;
+        marked = new boolean[g.getVertexCount()];
+        edgeTo = new int[g.getVertexCount()];
+    }
+
+    public boolean hasPathTo(int v) {
+        return marked[v];
+    }
+
+    public LinkedList<Integer> pathTo(int v) {
+        if (!hasPathTo(v)) {
+            return null;
+        }
+        LinkedList<Integer> stack = new LinkedList<>();
+        int vertex = v;
+        while (vertex != source) {
+            stack.push(vertex);
+            vertex = edgeTo[vertex];
+        }
+        return stack;
+    }
+}
